% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoWayANOVA.R
\docType{package}
\name{TwoWayANOVA}
\alias{TwoWayANOVA}
\title{Find test statistic value from a given data set and corresponding critical value}
\description{
This package finds test statistic value and corresponding critical value of parametric bootstrap based LRT, Min-T and Max-T tests for testing
equality of row effects against ordered alternatives in two-way ANOVA model with heteroscedasticity.
This package also finds simultaneous confidence interval for successive pairwise differences of treatment effects.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
